Group all swap encryption processes into a single 'Volatile' class
